Simplify the view property KVO tracking.
If things are changing and animating on screen, our outlines may get out of sync with the underlying view being tracked. Without tracking the entire view hierarchy, this will be hard to avoid. Previously we tried to catch some changes by observing the view’s superview, but since we don’t strongly retain the tracked view’s superview, it may go away while we’re observing it. Rather than trying to patch this incomplete solution, we’re simplifying and only observing the frames of the tracked views.