Add hack to FLEXRuntimeUtility to add a frame property on UIView.
Despite being declared as a property in UIView.h, frame is not a property from the perspective of the runtime. In order to easily use our property editing interface for viewing and changing view frames, we add the frame property to the UIView class at runtime if it doesn’t already exist.