First step refactor of FLEXArgumentInputView
We will now have a generic FLEXArgumentInputView superclass that just shows a title and declares the common interface for setting an initial value for the argument, getting the user’s input, etc.

With FLEXArgumentInputViewFactory, we will be able to pick different FLEXArgumentInputView subclasses depending on the type encoding of the argument. This will allow us to go beyond string-only input to have things like switches for BOOLs, multiple fields for structs, color pickers for UIColors, font pickers, etc.