Keep the view explorer’s shortcutPropertyNames internal for now.
We’re going to provide defaults internally for UIView and some commonly used UIView subclasses in UIKit. At some point we might consider a way to let users of the library customize the shortcut property list for their own classes. We might also consider moving this functionality out of the view explorer and up into the object explorer.