Reload the entire object explorer table on appearance.
We want to keep the data fresh, especially since a lot of the drill-ins can lead to changes in the data displayed in the table. Previously we reloaded only the visible cells, but this breaks when the changes alter the what we want to filter out with the search text (leads to a table view assertion). Now we’re reloading the entire table to make sure we’re safe and have the latest info displayed.