Added an API to register custom entries in the Globals table view controller.
This way 3rd party applications can extend the functionality of FLEX by adding quick access to important objects in their applications.
I modified FLEXGlobalsTableViewController to keep track of the items to display in the table with a list of FLEXGlobalsTableViewControllerEntry objects which have 2 blocks: one that gives you the name of the entry to display, and one that gives you the view controller to push when that entry is selected. I used a block to make the creation of both lazy because they can change (for example in case the application window changes).
I exposed this partially in the API: users can pass a reference to an object. This way they can set this app lazily in a +load method in some class, and not wait until the object they want to be able to inspect with FLEX is actually created, which simplifies the use (but this may be my brain being contaminated by ReactiveCocoa, so I'm definitely open to criticism!)

Open questions:
Should this API be in FLEXGlobalsTableViewController, which means that users of the API must #import that header, or should it be in FLEXManager?