Add copy support to description cell in FLEXObjectExplorerViewController
Right now the description cell in `FLEXObjectExplorerViewController` uses `FLEXSingleRowSection`, which doesn't have the copy support. Thus long press copying is not an option there.

The other cells are using `FLEXMetadataSection` which has good support. This change is going to add a basic "Copy Text" menu action for single row section.